Use purego to have much more code in common between unix and windows
